:hammer: Refactor: 이미지 리소스 저감 조치
