+++ src/components/Landing/Main/FirstSection/Navigation/Navigation.scss
@@ -12,7 +12,6 @@
   }
 
   & ul li {
-    min-width: 150px;
     margin: 10px 30px;
   }
 
@@ -23,7 +22,7 @@
   }
 
   & ul li img {
-    width: 45px;
+    width: 50px;
     height: 50px;
   }
