+++ src/components/Landing/Main/ForthSection/index.scss
@@ -14,8 +14,7 @@
 }
 
 .landing-forth-section-title {
-  // width: calc(50% - 50px);
-  width: 80vw;
+  width: calc(50% - 50px);
 }
 
 .landing-forth-section-text {
@@ -71,6 +70,10 @@
     align-items: center;
 
     max-width: 80vw;
+  }
+
+  .landing-forth-section-title {
+    width: 80vw;
   }
 
   .landing-forth-section-img {
