+++ src/components/Landing/Main/SecondSection/index.scss
@@ -9,7 +9,7 @@
   padding: 30px 0;
   min-height: 900px;
 
-  background-image: url('../../../../lib/assets/images/landing/section/section02/image_00.png');
+  background-image: url('../../../../lib/assets/images/landing/section/section02/image_00.webp');
   background-repeat: no-repeat;
   background-position: center;
   background-size: contain;
